cloudgovernance-master/azure-governance/src/requirements.txt
###### Requirements without Version Specifiers ######`
click
pandas
azure-cli
azure-mgmt-subscription


###### Requirements with Version Specifiers ######`